AGS: Common: fix String::Trim*() to avoid asserts for extended chars
This complements 82abea8
From upstream 82ffab07df0bdf4a016addcc6050ba7f2b5e8fdc

+++ engines/ags/shared/util/string.cpp
@@ -780,7 +780,7 @@ void String::TrimLeft(char c) {
 		if (c && t != c) {
 			break;
 		}
-		if (!c && !Common::isSpace(t)) {
+		if (!c && !Common::isSpace(static_cast<uint8_t>(t))) {
 			break;
 		}
 		trim_ptr++;
@@ -807,7 +807,7 @@ void String::TrimRight(char c) {
 		if (c && t != c) {
 			break;
 		}
-		if (!c && !Common::isSpace(t)) {
+		if (!c && !Common::isSpace(static_cast<uint8_t>(t))) {
 			break;
 		}
 		trim_ptr--;
